FAC-627: refactor - Minor Repackaging
