+++ flickit-advice-engine/src/test/java/org/flickit/assessment/advice/application/port/in/advice/CreateAdviceUseCaseParamTest.java
@@ -1,8 +1,7 @@
-package org.flickit.assessment.advice.application.port.in;
+package org.flickit.assessment.advice.application.port.in.advice;
 
 import jakarta.validation.ConstraintViolationException;
 import org.flickit.assessment.advice.application.domain.AttributeLevelTarget;
-import org.flickit.assessment.advice.application.port.in.advice.CreateAdviceUseCase;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
